Impromtu chaos threads

+++ bots/chaos/src/guilds/services/ChannelService/ChannelCleaner.ts
@@ -1,27 +1,25 @@
 import { HadesClient } from "@hades-ts/hades";
-import { TextChannel } from "discord.js";
+import { Channel, ChannelType, TextChannel } from "discord.js";
 import { inject } from "inversify";
-import { ConfigGuild } from "../../config";
-import { guildSingleton } from "../decorators";
-import { tokens } from "../tokens";
+import { ConfigGuild } from "../../../config";
+import { guildSingleton } from "../../decorators";
+import { tokens } from "../../tokens";
 import { DataService } from "./DataService";
 import { CleanupBypass } from "./CleanupBypass";
+import { WithRequired } from "../../../types";
 
 
 @guildSingleton()
 export class ChannelCleaner {
 
+    @inject(HadesClient)
+    private client!: HadesClient;
+
     @inject(tokens.GuildId)
     guildId!: string;
 
     @inject(tokens.GuildConfig)
-    private config!: ConfigGuild;
-
-    @inject(tokens.ChannelFactory)
-    private getChannel!: () => Promise<TextChannel>;
-
-    @inject(HadesClient)
-    private client!: HadesClient;
+    private config!: WithRequired<ConfigGuild, 'channel'>;
 
     @inject(CleanupBypass)
     private bypass!: CleanupBypass;
@@ -37,7 +35,23 @@ export class ChannelCleaner {
             return;
         }
         
-        const channel = await this.getChannel();
+        const channel = await this.client.channels.fetch(this.config.channel.id);
+        
+        if (channel === null) {
+            console.error(`[ChannelManager] Channel ${data.thread} does not exist for guild ${this.guildId}`);
+            return;
+        }
+
+        if (!('messages' in channel)) {
+            console.error(`[ChannelManager] Channel ${data.thread} is not a text channel for guild ${this.guildId}`);
+            return;
+        }
+
+        if (!('bulkDelete' in channel)) {
+            console.error(`[ChannelManager] Channel ${data.thread} does not support bulk deletion for guild ${this.guildId}`);
+            return;
+        }
+
         const messages = await channel.messages.fetch();
         
         await channel.bulkDelete(messages.filter(message => {
@@ -50,10 +64,10 @@ export class ChannelCleaner {
             }
 
             if (message.author.id === this.client.user?.id) {
-                if (this.config.singleMessage) {
+                if (this.config.channel.singleMessage) {
                     return true;
                 }
-            } else if (this.config.guardChannel) {
+            } else if (this.config.channel.guard) {
                 if (!this.bypass.isExempt(message.author.id)) {
                     return true;
                 }
